feat: refactor deleting_files.php to utilize HTML forms with POST methods to handle deleting and creating the file 'delete_file_test.php' to easily traverse directories and delete and create the file at the click of a button.

+++ LearnPHP/WorkingWithFiles/deleting_files.php
@@ -1,18 +1,49 @@
 <?php 
-//TO PROPERLY SEE FILE DELETION, YOU MUST BE ON THE 'deleting_files.php' PAGE IN THE BROWSER...
 
-//use unlink() method that takes in a single parameter, the file name as a string. unlink() will delete the file you specify...
-unlink("delete_file_test.php"); 
+  if(isset($_POST['delete_file'])) {
+    unlink("delete_file_test.php");
+  }
 
-//uncomment the following lines and refresh the browser to create the file to be deleted. After the file is created through the browser refresh comment back out the following lines and refresh the browser while on deleting_files.php
+  if(isset($_POST['create_file'])) {
+    $file = "delete_file_test.php";
 
-// $file = "delete_file_test.php";
-// if( $handle = fopen($file, 'w')) {
-//   fwrite($handle, "Hi, I am the file that is going to be deleted if you follow the instructions properly!");
-//   fclose($handle);
-// } else {
-//   echo "The application was not able to write on the file...";
-// }
+    if( $handle = fopen($file, 'w')) {
 
+      $delete_page_form = "<form action=\"deleting_files.php\">
+      <input type=\"submit\" value=\"Click me to go to deleting_files.php page!\">
+    </form>";
+
+      fwrite($handle, "Hi, I am the file that is going to be deleted if you follow the instructions properly!"."<br>".$delete_page_form);
+
+      fclose($handle);
+
+    } else {
+      echo "The application was not able to write on the file...";
+    }
+  }
 
 ?>
+
+
+<?php include "partials/header.php"; ?>
+
+  <!-- Form to move up a directory to see if the file was deleted successfully... -->
+  <form action="./">
+    <input type="submit" value="Click me to go up a directory!">
+  </form>
+
+  <!-- Form to Delete the file onClick -->
+  <form action="deleting_files.php" method="POST">
+
+    <input type="submit" name="delete_file" value="Click me to delete the file 'delete_file_test.php!'">
+  
+  </form>
+
+  <!-- Form to create the file to be deleted... -->
+  <form action="deleting_files.php" method="POST">
+
+    <input type="submit" name="create_file" value="Click me to create the file to be deleted!">
+  
+  </form>
+
+<?php include "partials/footer.php"; ?>
